Moved modals & styled modals to be their own component
